Resume-BackEnd/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
scikit-learn
numpy
pandas